% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE143371_count.R
\docType{data}
\name{GSE143371_count}
\alias{GSE143371_count}
\title{Example dataset of gene expression matrix}
\format{
A matrix with 57131 rows and 6 columns:
}
\source{
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE143371
}
\usage{
data(GSE143371_count)
}
\description{
This dataset contains normalized expression values
for demonstration purposes.
Normalized process was performed by DESeq2 as follows:
}
\details{
dds <- DESeqDataSetFromMatrix(
	countData=countData,
	colData=sample_info,
	design=~factor(level))
rld <- rlog(dds, blind=FALSE)
GSE143371_count <- assay(rld)
}
\keyword{datasets}
