% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE143371_group.R
\docType{data}
\name{GSE143371_group}
\alias{GSE143371_group}
\title{Example dataset of group labels}
\format{
A vector with 6 elements (three "Undiff" and three "Neu"):
}
\source{
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE143371
}
\usage{
data(GSE143371_group)
}
\description{
This dataset contains group labels for GSE143371_count.
}
\keyword{datasets}
