% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveNetworkData.R
\name{retrieveNetworkData}
\alias{retrieveNetworkData}
\title{Retrieval of TF - target gene table and TF metadata from ChIP-Atlas}
\usage{
retrieveNetworkData(assembly = c("hg19", "mm10"), out.dir = tempdir())
}
\arguments{
\item{assembly}{Genome Assembly (hg19 or mm10)}

\item{out.dir}{Output Directory (Default: tempdir())}
}
\value{
A corresponding table between TF (SRX ID) and target gene (Gene Symbol) and TF metadata (SRX_ID, TF, celltype1, celltype2)
}
\description{
Two tab-delimited files are retrieved from ChIP-Atlas.
To check the local cache path, type the code below:
tools::R_user_dir("BiocFileCache", which = "cache")
}
\examples{
library("wPGSAR")
network_data <- retrieveNetworkData()
}
