% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wPGSA.R
\name{wPGSA}
\alias{wPGSA}
\title{Weighted Parametric Gene Set Analysis (wPGSA)}
\usage{
wPGSA(network_data, expression_data, group = NULL)
}
\arguments{
\item{network_data}{Output object of retrieveNetworkData()}

\item{expression_data}{Gene expression matrix specified by user}

\item{group}{Group label vector to perform statistical tests to compare TF activity between groups. (optional)}
}
\value{
A list containing the result of wPGSA:
\describe{
  \item{tstat}{TF activity scores (t-statistics) computed from weighted expression and background.}
  \item{tstat_se}{Standard error used in tstat calculation.}
  \item{group_mean}{Group-wise averages of t-statistics, estimated by linear regression.}
  \item{group_se}{Standard errors of group means.}
  \item{group_t}{t-values of group means (group_mean / group_se).}
  \item{pval}{P-values from linear regression testing whether group means differ significantly from 0.}
  \item{fdr}{Adjusted p-values (FDR) using Benjamini-Hochberg correction.}
}
}
\description{
Weighted Parametric Gene Set Analysis (wPGSA) to infer the activity of transcription factors by evaluating the expression variance of their target genes using weighted t-statistics. When group labels are provided, the package can also perform statistical tests to compare transcription factor activity between groups.
}
\examples{
library("wPGSAR")
data(GSE143371_count)
data(GSE143371_group)
network_data <- retrieveNetworkData()
res1 <- wPGSA(network_data, GSE143371_count)
res2 <- wPGSA(network_data, GSE143371_count, GSE143371_group)
}
